t5000-valgrind: run multi-broker instance
Ensure at least some code for rank > 0 is run in the valgrind test
by executing multiple brokers instead of just a single broker in
t5000-valgrind.t.

Default size of test is 2 brokers, may be overidden by
VALGRIND_NBROKERS.

In the case of running many brokers, the shutdown grace period may
need to be increased as well, so allow a VALGRIND_SHUTDOWN_GRACE
environment variable to do that (default=16s).