kvs: Rename 'namespace' variable globally
The keyword 'namespace' is a reserved word in C++, so globally
remove use of the variable name 'namespace'.

Generally speaking, all local variables and function parameters
that were 'namespace' are now 'ns'.  In a few locations that
'namespace' was previously changed to 'name_space', use 'ns'
instead.

Structs that stored a namespace name now use the variable 'ns_name'.

In kvs-watch, rename 'struct namespace' to 'struct ns_monitor', and
use the variable 'nsm' for that struct.

Fixes #1963