cmd/builtin/hwloc: add aggregate-load plumbing subcommand
Add an undocument `aggregate-load` subcommand to flux-hwloc. This
subcommand is meant to be run across all ranks to aggregate a
JSON topology summary and optionally store it in the KVS. The
subcommand also optionally loads local hwloc topology XML into
the appropriate resource.hwloc.xml directory for the local rank.