modules/job-manager: set FLUX_JOB_CANCELED on reload
Problem: if job-manager module is reloaded, canceled
jobs (that have not been --purged) return to the queue.

If a job's event log has the 'cancel' event, then
set the FLUX_JOB_CANCELED flag on the job when
reloading job state from the KVS.  Then, at least for now,
if a job with that bit set is retrieved during reload,
don't add it to the queue.