libjob: add RFC 21 job states
Define the job states proposed in RFC 21 in job.h.
Use an enum, but set bit values to allow a mask of
job states to be passed to a yet-to-be-implemented
flux_job_wait() interface.

Add a job state field to the job data structure
used to track jobs in the job manager, and include
the job state in the job listing output.

Adjust flux job list column numbers in job manager
sharness test to account for new state column.

In this commit, the state is simply set to NEW
and that's it.