configure: check for asciidoctor version >= 1.5.7
Problem: asciidoctor versions < 1.5.7 fail to properly generate
multiple manpages from the same adoc source, and so should not be
used for doc generation.

Generate a warning and clear the ADOC autoconf variable when a2x is
not found, but an older asciidoctor is found, so that manpage
generation is disabled. Issue a valid warning to the user both
at the time the asciidoctor version is checked as well as when
configure completes.

Fixes #1672
Closes #1911