job-ingest: make max count optional in default jsonschema
Problem: rfc14 was recently updated to allow `min` without `max` to
be specified to indicate "infinite" maximum count. Some examples
were updated that would now fail the default jsonschema validator
in flux-core.

Update the schema to have "min" be the only required property of the
"complex_range" specification, but if any one of "max", "operator",
or "operand" are present, all must be.