modules/kvs: Remove watchlist & add kvssync
As kvs watchers no longer exist in the kvs module, remove the old
kvs watchlist code that still existed due to its use within kvs.sync.

Replace the kvs.sync mechanism with a new internal lib called "kvssync"
that implements a very simple sync mechanism.  All sync-waiters are added
to a sorted list on kvsroot.  When a setroot() has occurred, responses
are only sent to those waiters at the front of the list.