codecov.yml: adjust codecov/patch threshold
Problem: There is some "noise" in Codecov.io coverage reports as
run via Travis-CI (possibly due to running tests in parallel?), and
PRs that do not even affect C code may affect coverage of the
project as a whole.

To reduce false positives, adjust the "failing" codecov threshold
from the defaul to 0.1%, which should give some buffer for noise
in the coverage test results.