codecov.yml: adjust range, always post new report
Adjust codecov.io range from 50-100 to 65-100 so that 65% coverage
shows as red in reports.

Additionally, it is confusing that codecov.io updates its old post
instead of posting a new report when a branch is updated. Change
the comment behavior to "new" so Codecov deletes the old comment
and posts a new one on each run.