modules/job-manager: publish job-exception event
Problem: An event named 'job-cancel' is published when a
job is cancelled, but cancellation is just one of several
possible exceptions that can abort a job.

Rename the event that published when a job is canceled to
'job-exception' and include exception severity and type.

There are no subscribers to this event yet, but the intent
was that it would inform dependency service, scheduler,
and exec service that any pending requests from this job
should be aborted.