cmd/flux-kvs: Support -O and -s on write ops
Support a new -O and -s option on write operations like put,
unlink, link, and mkdir.  The option will output the RFC11 root
treeobj or root sequence number upon a successful write.
These outputs can be used for causal consistency, such as by feeding
the treeobj to a lookup-at or the sequence number to a wait version.