job-manager: allow a job to be placed in multiple queues
Generalize 'struct queue' to prepare the way for
other queues to exist in the job manager.

The 'struct queue' lookup hash is now optional.
Since all active jobs will be in the main queue with
the hash to speed up lookup by id, other queues
containing a subset of active jobs can leverage the
main queue's lookup hash and thus don't need one.

The "list handle" is exposed in the API so that
users can assign to the appropriate struct job
member (or ignore it).  Add job->aux_queue_handle,
so that a job can be be in another queue besides
the main one.  Note the handle is returned by insert()
and is used as input to delete() or reorder().

Change the queue implementation so that the job
reference count is incremented when inserted into
the list and decremented when removed from the list.
If the optional hash is enabled, it tracks the list
(not taking a reference on jobs) not the other way
around like before.

Update users and tests.