job-manager: add scheduler interface
Add the scheduler interface, defining an initial
handshake so that the scheduler can be unloaded
and reloaded on a live system while still accounting
for allocated resources.

Implement two scheduler modes:

single: only one alloc request is allowed to be
outstanding to the scheduler at one time.  This is
appropriate for the FCFS scheduler, and requires
that the job manager implement a queue of jobs
that are ready to allocate resources.

unlimited: an alloc request is immediatley sent
when a job becomes ready to be allocated.  The
queue is actually still useful in the event that
the scheduler is not loaded when jobs become
ready.

Define alloc and free RPCs that do not use matchtags
or futures.  One side effect is that a "failure"
returned in the usual RFC 6 way cannot be associated
wtih a job, so we have to encode the expected forms
of failure in a "success" response payload, and
reserve the normal failure for disasters that should
stop the scheduler interface.

The code to advance the job state and log eventlog
messages leaves a bit to be desired here, but it's
a start that can be improved upon later.