modules/job-manager: cancel advances state to CLEANUP
When a job is canceled, change its state to CLEANUP.
At this point, since we can have no other work pending,
CLEANUP should move the job to the inactive directory
and advance the state to INACTIVE.  However, for now
just leave the job in the CLEANUP state and do not
remove it from queue.

Rename the internal FLUX_JOB_CANCELED flag to
JOB_EXCEPTION_PENDING and clear it once event is logged
to eventlog.  Since state is not advanced to CLEANUP until
the exception event is logged to the eventlog, this flag
may be useful to prevent races.

Change the restart logic so that a canceled job is
reloaded in CLEANUP state.

Update job manager sharness test to ignore jobs in
CLEANUP state where appropriate.