job-manager: handle partial submit failure
Problem: jobs are submitted in batches to the job manager.
If one job fails to be added, all the jobs fail.  However,
when there is a failure, some jobs may be left in the queue
after their users were told they failed.

Keep a list of newly added jobs.  On failure, walk the
list and dequeue the jobs.