job-manager: wire in scheduler
When the ingest module adds jobs to the queue,
advance the state to SCHED (skipping DEPEND
because we don't have that yet) and insert the
job in the "alloc_ctx" queue.  Once the scheduler
is ready, a sched.alloc request will be sent
on behalf of the job.

When the job-manager is restarting from the KVS,
do the same thing for newly read-in jobs in the
SCHED state.