modules/job-ingest: add context to submit event
Problem: userid and priority are written to separate
keys in the KVS job schema.  job-manager has to read
these to recreate job state when it restarts.

Add userid and priority to job eventlog "submit"
event, so that the job manager can get info when it
replays the eventlog and won't need to access the
other keys.