modules/job-manager: add new job constructor
Problem: the code for recreating the queue from
job data in the KVS is quickly becoming unreadable.

Add job_create_from_eventlog(), which creates a
a struct job by replaying a snapshot of the job's
KVS eventlog.

RFC 21 documents that the job state can be
recreated by replaying the eventlog. This function
implies that this property extends to the whole
in-memory job-manager state for a job.

This depends on the "submit" event including the
userid and initial priority, just added in an
earlier commit to job-ingest.  If those are missing,
creation fails.