travis-ci: unshallow git repo before running docker
On travis-ci the version of git(1) inside of our docker image may
be too old for the `git fetch --unshallow --tags` to succeed, as
evidenced by errors like:

 git fetch --unshallow --tags
 fatal: unknown value for config 'protocol.version': 2

in our travis builds. To ensure we don't hit this error, run
the command to fetch tags *before* running docker, using the version
of git installed in travis.

Fixes #2024