libschedutil: add scheduler helper functions
We will initially have sched-dummy for testing,
and, shortly, sched-simple for our first functional
in-core scheduler.

There is significant boiler plate code needed
to handle the sched - job-manager interaction
that would be duplicated in both schedulers, so
rather than cut & paste, put it in an internal
library, "libschedutil".

After we've had some experience with it, hopefully
it can be improved to the point that it can be
exported as part of the API for external schedulers
like the flux-sched project.