flux-hwloc: add cpuset to resource.hwloc.by_rank
Problem: the resource.hwloc.by_rank key encodes the number of cores
available on each rank, but does not encode the logical core ids, and
not all available core lists will be strictly core0-core(n-1).

Add a new `cpuset` key to the by_rank JSON object which encodes the
list of available cores for the indicated rank, so that this list
may be used by readers of this object.