testsuite: t0015-cron.t: improve reliability
Problem: there is an inherent race in the cron "min-interval" test,
which ensures that a cron "event" entry will not run more than
once per specified minimum interval. In the test, there is no way to
guarantee that the check before the sleep runs before the interval has
expired, and thus on busy or slow systems the test may fail because
the stats.count is >1 instead of 1.

Since there is no way to guarantee that our test command runs within
the minimum interval specified, dump the idea of testing this way,
and instead ensure the cron job was delayed by checking for a log
message from the cron service.