testsuite: double timeout for test_under_flux test
The t0001-basic.t 'test_under_flux' test is unreliable on slow or
loaded machines. It might be that the 5s timeout is not quite long
enough to get the underlying flux instance being spawned by
test_under_flux instantiated and terminated.

Double the timeout to 10s in hopes the test becomes more reliable.