libidset: optimize idset_count
Problem: idset_count(3) needs to iterate the entire vEB tree to
get a count of entries in the set on each call, and this is
(relatively) slow. Considering use cases of idset_count(3) such
as in a sort algorithm, an improvement here may be warranted.

Add a count member to the idset structure and keep the embeeded
count up to date within the idset_set/clear set of calls.

This unfortunately adds some slight overhead to the set and clear
operations, since each integer argument to set/clear has to
first be checked for membership (or lack thereof) in the idset
so that the count can be appropriately updated.