job-manager: add event batcher
Problem: several subsystems in the job-manager independently
commit events to the job KVS eventlog, duplicating code and
missing an opportunity to batch temporally close commits
together for reduced overhead.

Add an new eventlog update mechanism, similar to the KVS update
batching in job-ingest.  Accept updates via two functions:
event_log(), and a varargs version, event_log_fmt(). These
functions add to a KVS transaction that grows in memory until
a timer expires, then the transaction is committed in one RPC.

Callbacks can optionally be registered with each event.
The callbacks are called upon completion of the transaction,
with access to the future used for the commit, so they can
determine whether it succeeded or failed.

A hardwired tunable, initially set to 10ms, determines how long
a batched transaction waits until it is flushed.  This value
may be tuned as we gain experience with the system.

If job-manager is trying to unload with commits in flight, block
until those complete, and invoke any registered callbacks before
tearing down.