job-manager: make event_log commit failure fatal
Problem: event_log() callback is needed even if all
it does is log the error.

Change the event_log() KVS commit continuation to
log an error and stop the reactor on failure, so
that a callback is only needed if the user needs
to be notified when the event has been committed.

Drop event_log callbacks in alloc, priority, and
raise source modules.  In priority and raise,
this changes the control flow so that a response
may be received before the eventlog update is
finalized, which affects the timing of some tests.