job-manager: add exec service interface
Implement exec service protocol in start.c (see
documentation of protocol at the top) as discussed
in #2040.  The protocol strongly resembles the
sched protocol, eschewing matchtags and futures,
and requiring jobid's to be carried in all responses.

The design allows the exec service to be loaded after
the job-manager, and to be overridden to support
the simulator initial program that provides a simulator-
specific exec service.

The exec service may only be (re-)regstered on an idle
system, that is, one with no outstanding start requests.

Wire the exec service events into the job state machine.

Change job-manager job unit test to insert a severity=0
exception in submit+alloc+free test input, since state
machine now requires free to be issued from CLEANUP state.