testsuite: enhance valgrind/job workload
Now that a scheduler and dummy execution module are loaded
during the valgrind `job` workload, ensure the test waits for
all jobs to run by adding a call to `flux job wait-event` to
the end of the script.

Ensure jobs run in order for now by limiting resources to a
single core.

Add test for cancelation of a job, to ensure no leaks there.

Finally, simplify the script by using `flux jobspec` jobspec
generator tool, instead of hard-coded jobspec string.