job-exec: add prototype job-exec module
Add do-nothing job-exec service that runs through exec system protocol
by sleeping for duration specified in system.attributes.duration
of the jobspec. If a duration is not specified the job "finishes"
almost immediately (10us).

An object in jobspec, attributes.system.exec.test, can be used to
set other test parameters on a per-job basis (see documentation
in the code for more information), including forcing the module
to throw an exception during initialization or run of the job.