job-manager: add CLEANUP->INACTIVE transition
Generate a 'clean' event once activities in CLEANUP
have concluded.  The 'clean' then event triggers the
transtion to INACTIVE state.  It can also serve as an
EOF marker in the job eventlog.

Add an action for the INACTIVE state: removal from the queue.
It is left for a future PR to move INACTIVE jobs from the KVS
job.active -> job.inactive directory.

Update job-manager-dummysched sharness test:  if the job is
not in the queue listing, but has the 'clean' event in its
eventlog, report state as I.