t/job: Update tests for RFC21 eventlog changes
Strings in the default flux job eventlog output are now quoted, so adjust
tests that previously did not assume quotes around strings.

Additional notes in an event are now listed via a key of "note" and
not just additional text at the end of a line.  Update tests
accordingly to check for "note=".