job-ingest: allow textual error from job-manager
Problem: job-manager may refuse a submit request if
it is being drained, but its explanatory error message
is not returned to the user.

Change the logic to ensure that the human readable error
is returned in the ingest submit response.

Also disable logging at LOG_ERR level when submit fails,
since the drain error may now be expected in normal operation.