libflux: propagate child errors to composite future
Problem: Composite wait "any" or "all" futures are fulfilled with
sucess even if one of the child futures that makes up the composite
is fulfilled with an error. This makes it tricky to use the composites
as a component in a flux_future_and_then() chain, since the error
will not break the chain. This can cause unexpected failures from
the next continuation in the chain.

When fulfilling the dummy "parent" future for a composite, propagate
an error if a child future was fulfilled with an error. This means
that a composite "wait all" future is only fulfilled with success
if all of its child futures are fulfilled with success, and that a
"wait any" is fulfilled with success if the first future to complete
is sucessful.