configure.ac: fix asciidoctor version check
Problem: The check for asciidoctor version >= 1.5.7 was flawed in
some way, and failed to grab the version in some or all circumstances.
This caused the build of docs to be silently disabled in Travis-CI
for example.

Simplify the sed(1) expression used to grab the asciidoctor version
for `asciidoctor --version` output (tested working). Furthermore,
add a surrounding AC_MSG_CHECKING/AC_MSG_RESULT block so that the
detected version is printed in output for aid in future debugging
efforts.

Fixes #2102