t2401-job-exec-hello.t: fix incorrect use of run_timeout
In the test for job-manager.exec-hello failure when jobs still exist
with outstanding start requests, a `run_timeout` was used to execute
a `test_must_fail` shell function, which clearly cannot work.

Since we still want to be able to time out the script that is trying
to register with the job-manager (since a failure would hang forever),
use `test_expect_code 1` to differentiate between a timeout and
an expected failure.