job-manager: add job-state event publishing
Add a JSON array to the KVS batch machinery in event.c
that tracks [jobid, state] tuples.  When the post function
handles an event that triggers a job state change, add
the job and its new state to this object.

Once the triggering event(s) are committed to the KVS,
publish a job-state event message, notifying interested
parties that a set of jobs have transitioned to a new state.
The fact that this comes after the KVS commit has completed
maintains the invariant from RFC 21 that notification of
certain state changes indicate that info, such as exception
events or exec finish events, is fetchable from the KVS.

The timed batch machinery, already there to mitigate
the load on the KVS from bursts of job events, now
helps reduce the number of published event messages.
Each message will contain more job ID's, and the latency
will increase a bit from event occurrence to notification.