job-manager: drop cb, arg from event_job_post()
event_job_post() and event_job_post_pack() both include
a callback mechanism, so that the caller can be informed
upon completion of the batched KVS commit of eventlog entries.
This was originally intended for both error handling and
as a way to usefully tie state notification to KVS updates.

The errors are now handled directly in the callback.
We will generate state notification events directly as
well, so it is best for simplicity to remove this mechanism
that is otherwise not used.