libflux: drop errnum from flux_respond()
Problem: flux_respond() can return errors as well as
success, which sometimes leads to awkward and unsafe
code, such as code that depends on errno being set to zero,
when it is undefined *unless* there is an error.

Drop errnum from flux_respond().  Now it should only
be used to return success.

Update tests and users.

Fixes #2089