don't retry failed response
Problem: When flux_respond_*() fails in a service's
request handler, the error should not be handled by
calling flux_respond_error(), since that is just as
likely to fail as the first call.

This unlikely error will probably be the result of a
coding error, or a fatal error on the flux_t handle.
Reasonable coping mechanisms are stopping the reactor,
and/or by logging it at LOG_ERR level.

This pattern occurs in the broker and the KVS.
Log the error instead of retrying.