rework response error handling
Problem: the new footprint of flux_respond() and
flux_respond_error() favor a pattern for response
handlers that separates error and success code paths.
The error path ends in flux_respond_error() and the
success path ends in flux_respond().  However, some
services use a different pattern that combines error
and success code paths.  This pattern is now a poor
fit for the respond functions.

Replace this pattern in the content-cache, the broker
log service, kvs, sched-simple, cron, and aggregator.