broker: derive per-broker rundir from global rundir attribute
Change the user settable `broker.rundir` attribute to a global
`rundir`, presumed to be shared per instance if set. Add a new
`broker.rundir` which is now set to `${rundir}/rank` to allow sharing
of the global rundir.

The broker.rundir is still used as before, as the basis for ipc socket
paths as well as the connector-local socket and basis for FLUX_URI.

This change allows a global `rundir` to be set for a flux instance,
from which the FLUX_URI for any rank may be derived.

In flux-start, there is now no need to set a per-broker broker.rundir,
since this is now handled internally. Clean up the set `rundir`
recursively since there will be one subdirectory per launched broker.