testsuite: add test prog for arbitrary rpc payload
Add a test helper that sends a single RPC, using topic
string from command line and payload from stdin.  Any
response payload is displayed on stdout.

$ jq -j -c -n  "{name:\"rank\"}" | t/request/rpc attr.get

If a numerical argument follows the topic string, it indicates
that a failure expected with errno == the number.  In that case,
a failure response with the expected errno is treated as success,
and a success response, or a failure response with the wrong
errno is treated as failure.