libflux: add flux_future_fulfill_with()
Add an implementation of flux_future_fulfill_with(3), which allows
one fulfilled future to be used to fulfill another.

This function not only copies the fatal error, error result, or nornal
result from the source future into the destination future, but "embeds"
the source into the destination future such that the source future's
aux items are available read-only in the destination.

This makes it possible to use one future to fulfill another even if
some get() methods used on that future require auxillary data stored
within the object. (e.g. flux_kvs_lookup_get(3)).

Since flux_future_fulfill_with() propagates a copy of the result from
the source future to the destination future, the function may actually
be called multiple times, as long as the source future is reset
between calls.

On the first call, however, the source future is embedded in the
destination future, so it is an error to call flux_future_fulfill_with()
multiple times on the same target future with a different source.

This function steals a reference to the source future, since its
destruction has to be tied to the destination.