cmd/flux: Fix unlink corner cases with links
Fix corner cases in kvs unlink in regards to links.

If the link pointed to an invalid target, namespace, or
had an infinite cycle, the unlink_safety_check() function
would fail and not allow the user to remove the link.

Instead, unlink should always allow links to be removed.
Re-work unlink_saftety_check() to ensure links are always
removable.

Add regression tests.

Fixes #2129