modules/job-exec: fix uninit jobid in EPROTO response
Problem: when a malformed start request is received,
a failure response is sent that includes an uninitialized
job id, since the id was unable to be parsed from the
request:

{ "id":0,
  "type":"exception",
  "data":{
    "severity":0,
    "type":"exec",
    "note":"job start failure: Protocol error"
  }
}

Skip sending this response.  An EPROTO response is also
sent, which causes the job manager to stop sending requests
to the exec service.  Add a comment about the non-obvious
purpose of this response.

Now a test can be written in a straightforward way to
ensure that the EPROTO is returned.