libflux: auto-continue chained futures when needed
Add a new 'continued' flag for chained futures to allow detection
of whether a future was continued from one of the and_then or
or_then callbacks. If a future was not continued; either because
the user failed to call flux_future_continue(3) in the callback,
or because the appropriate and_then() or or_then() callback didn't
exist; then auto-fulfill the next future in the chain with the
currently fulfilled future.

If a user's callback was *not* called, be sure to still destroy the
future 'prev' from the internal continuation to avoid leaking
a future in a branch where the user has not registered a callback
from which to destroy it.

Fixes #2135