python: prevent destruction of flux handle passed to mod_main
The pymod module does not own the memory for the handle passed to it
by the broker, and during module unload the hanlde was being closed
twice: once by the Python GC and once by the broker.

To fix, take an explicit reference in python to the handle before
giving it to the loaded python module.