libflux/message: refactor PROTO codec
Problem: There is duplicate code for accessing u32 values in
PROTO block, and the assignment of fields depending on message
type is confusing.

Combine rolemask, userid, bigint and bigint2 into a "u32 array",
with common accessors.

Add convenience macros mapping field names to positions in the array.

This will make changing PROTO easier in the future.