connectors/ssh: warn on stderr if ssh cannot be run
Problem: If the default path to ssh cannot be found
by the ssh connector, flux_open() only gets ENOENT
back in errno, which isn't much to go on.

Emit the failed path and a hint about overriding by
setting FLUX_SSH on stderr if the popen() of ssh
fails.

Fixes #2140