libflux/handle: add flux_matchtag_group()
Problem: we need a way to test if a matchtag belongs
to an mrpc or a regular rpc to decide whether an orphaned
one can be retired, but there is no way to ask what type
it is because details are hidden in the tagpool "class".

Add tagpool_group(), wrapped with flux_matchtag_group().

Add unit tests.