libflux/msg_handler: handle late response
Problem: if an RPC future is destroyed before its
last response is received, the RPC code leaks the matchtag.
If the response is later received, no attempt is made
to reclaim the matchtag.

Look at unclaimed response messages before dropping them
and see if it contains a live matchtag.  If we have enough
info in the response to determine if it is the last one,
free the matchtag.