streaming services: check FLUX_MSGFLAG_STREAMING
Problem: if a request is sent to a "streaming" RPC
service without the FLUX_MSGFLAG_STREAMING flag,
an orphaned, non-error response might prematurely trigger
matchtag reclaimation.

Services that send multiple responses should check for
the flag early (before sending non-error responses),
and respond with an EPROTO error response if the check fails.

An error response terminates both streaming and non-streaming
RPCs, so the fact that the STREAMING flag is not set in the
EPROTO response causes no confusion.

Add this check to the following services:
- rpc unit test (rpctest.multi)
- job-info.eventlog-watch
- kvs-watch.lookup (with FLUX_KVS_WATCH only)