python: fix future_then callbacks on streaming RPCs
If the same future has its callback executed (e.g., in the case of a
streaming RPC), an error occurs. After the first execution of the
callback, the bindings remove the python-side reference to the data
handle. So by the time that the second call happens, the data handle has
already been garbage collected.

Add a reference count dictionary to keep futures with a pending `then` callback
alive, even if there are no remaining references to the future in the user's
program scope. When a callback is first set on the future, add an entry to the
dictionary with the future itself as the key and a value of 1. Whenever a
future's callback is run, decrement the count in the dictionary for that
future, and whenever reset is called on a future, increment the counter. If
the counter hits 0, delete the reference to the future from the dictionary.