python: job.submit_async returns a Future and get_id uses wait_for
Wrap the return of `submit_async` in a python Future object rather than
return an opaque, c future pointer so that users can leverage all python
Future methods.

Add a `wait_for` to ensure that the future is fulfilled before getting
the id.  Otherwise, an error occurs if the submission future is not
fulfilled when `submit_get_id` is called.