libpmi2: drop incomplete PMI-2 library
Problem: our PMI-2 library is incomplete and is an
attractive nuisance at this point.

It was added to allow OpenMPI jobs to be launched under
flux when OpenMPI, through its MCA plugins, determines that
it should dlopen slurm's libpmi2.so.  OpenMPI could be
tricked via LD_LIBRARY_PATH into opening a Flux libpmi2.so
that provides the minimal API needed by OpenMPI, and maps
it to the simple PMI-1 wire protocol.  For more background
see flux-framework/flux-core#746

Since we submitted a flux MCA module that helps OpenMPI
choose the right PMI in the flux environment, and shipping
this libpmi2.so with flux is misleading as it doesn't implement
the full PMI-2 "spec", let's get rid of it.

Fixes #2082