libpmi: drop singleton and dlopen methods
Problem: the PMI implementation in libpmi is a source of
confusion because it's used by the flux broker to bootstrap
in non-Flux environments, AND it is the PMI library provided
to users for launching under Flux.

We can significantly dumb this thing down if it is assumed
to only be used by users running under Flux and then it will
be more understandable when looked at only from that perspective.
Drop singleton and dlopen (wrap) methods, and in fact drop the
method abstraction and just assume the "simple client" is the
only method.

Update pmi unit tests.