libpmi: move clique emulation to simple_client
Problem: the code for obtaining the "clique" information
from the 'PMI_process_mapping' key hardwires canonical
PMI API functions, but a user (internal for now) might want
to use the handled simple_client functions instead.

Move this code to the simple_client class.  Currently this is
used behind the canonical clique functions in the Flux libpmi.so,
but not internally.  In the past it was used to assist with
PGM wireum.

Update test/pminfo.