broker: Move assert in content_cache_destroy
Asserts added in

f8521c0ca734f3417cadeaa0ee1871fcbb81297c

were added to check that cache entry is removed only when there
is no pending load or store in the cache entry.  However, the
asserts are checked during broker teardown, which may tear
down the broker during emergency situations, leading to an assert
that may not be warranted.

To solve this, move the asserts from cache_entry_destroy() to
remove_entry(), so the asserts are only checked during normal
cache entry removal operations.

Replace the asserts in cache_entry_destroy() with log messages
indicating a load/store of a cache entry did not complete.

Fixes #1267