cmd/flux-srun: add simple run command
Add a simple run comand that wires together:
- flux jobspec srun
- flux job submit
- flux job attach

The idea is that this simple interface can be
back-ported to 0.11 so that users can change their
workflow scripts now with some guarantee that they
won't break when we finalize our more sophisticated
porcelain run command.

Despite the name, the plan is not to be a faithful
replica of the SLURM srun command, just something
simple and familiar that leverages the work Stephen
already invested in the flux jobspec command.
Thus, only the options supported by flux jobspec srun
are supported here.