build: break PMI server out of installed libpmi.so
Problem: ldd shows installed libpmi.so.0.0.0 depends
on libczmq, libzmq, and their dependedencies, but
libpmi was designed to be standalone.

These dependencies are due to inclusion of
simple_server.[ch] in the "noinst" internal libpmi.la,
but the simple_server code is for internal use only and
is not needed in the installed client-only library.

In libpmi, build two noinst libraries: libpmi_client.la
and libpmi_server.la, and only include the former in
the installed libpmi.so.