testsuite: add coverage for unimplemented PMI wire proto
Problem: there is no test coverage for server-side
interpretation of publish, unpublish, lookup, and spawn
requests.

These requests all return PMI_FAIL and are not fully
implemented on the server, but we should at least exercise
the code and make sure it gets the response to the user.

Add dummy client code to the test and make each request.

Enable tracing on the server side and route trace telemetry
through libtap diag(), then remove a few explicit diag() calls
that are now redundant.