broker: move register_attrs calls
Move all "register_attrs" calls to a location closer to the
initialization / setup of other parts of the object.  With this
change, the register_attrs calls are more in line with their
object and removes the impression that the registering of the
attrs needs to be done in a specific order except for when
commented.

The movement of content_cache_register_attrs() effectively
reverses commit

3d59fbf8c2343b40b0f8638d02c7ab3cc9d8fd6a

which was a fix for #1063.  The fix of issue #1036 no longer
requires the content_cache_register_attrs() call to be made
far earlier in the code before zsys_init().