t2500-job-attach: set all jobs to use testexec exec implementation
Tests of flux-job attach should use testexec execution implementation
(for now at least) so that jobs run for set timelimit and behave in
an expected way.

Check for jq(1) during test startup, and if preset add the appropriate
key to emitted jobspec before submission to activate test execution.
This means that these tests are now skipped if jq is not present.