modules/job-exec: set FLUX_KVS_NAMESPACE for shell
Problem: the job shell may not be running as the
instance owner in the future, thus should redirect
all KVS operations to the guest namespace.

Set FLUX_KVS_NAMESPACE in the shell's environment
so all KVS accesses by the shell and the job will be
rooted in the guest namespace.

Update dummy job shell to verify that FLUX_KVS_NAMESPACE
is set, and skip setting KVSDIR and NAMESPACE.

Update one sharness test that depended on the job shell
writing to the primary namespace so that it now gathers
output from the guest namespace.