modules/barrier: [cleanup] make timer per-barrier
Problem: there is a single timer for the barrier module,
started when a barrier is entered (if not yet armed),
cleared when all counts have been sent upstream.  The
purpose is to batch up multiple barrier counts to send
upstream at once.  Sharing the timer between multiple
barriers is confusing and causes unpredictable latencies.

Create a per-barrier timer instead.  Now when multiple
barriers are active, their timings are independent.