check for invalid FP types in fsd
Adding a check to disallow NaNs and infinites in flux standard
date format parsing and serialization in fsd.  This came up while
looking into #2213.

Also added a test for EINVAL on "NaNs" and "infinites", this
also covers the various other forms of those, but I didn't see a reason
to test strtod.