docker: add valgrind-devel to centos7 Dockerfile
Problem: without the valgrind-devel package, valgrind.h is not found
during configure and the valgrind test is disabled by default. When
run with -d, the t5000-valgrind.t is forced to run, but may encounter
false positives and bad stack traces due to modules that are dlclosed.

Add valgrind-devel to the centos7 docker image. (valgrind.h is part of
the base valgrind package in Ubuntu)